+++ src/components/Text/HighlightedTitle.tsx
@@ -32,7 +32,7 @@ const HighlightedTitle: FC<TitleProps> = ({
         />
         <p
           className={clsx(
-            'relative z-10 whitespace-nowrap font-medium',
+            'relative z-10 whitespace-nowrap !font-medium',
             { 'header-2 md:header-1': size == 'large' },
             { 'header-3 md:header-2': size == 'medium' },
           )}
