+++ src/components/card/ChartDetailsCard.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 
-const DataCard: FC<{ children: React.ReactNode }> = ({ children }) => {
+const ChartDetailsCard: FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
     <div className="body-normal rounded-xl bg-[#F4F4F4] p-4 leading-snug">
       {children}
@@ -8,4 +8,4 @@ const DataCard: FC<{ children: React.ReactNode }> = ({ children }) => {
   )
 }
 
-export default DataCard
+export default ChartDetailsCard
